[Tutorial] Ray Casting
It has been a while since I posted anything. A year. Well, technically
not a year, but the joke is already 5 months old, so we'll roll with it.

Studying back up on ray casting has been quite an experience. One thing
that I can say with confidence is that the tutorial has been tested and
the things that needed changing have been changed, to make sure there
are no mistakes.

Other uncaught mistakes should be commented at the bottom, or raised as
a Github issue in this repo.